Fixed and working version for bisphosphines
Hard coded and brittle, but a starting point.  Here is the associated content from the email:

Starting from geometry optimization: There are a number of scripts floating around to convert gamess/gaussian/qchem outputs to energies/xyz structures.  If those don't suffice I point you to open babel for other types of file / type conversion that might be amenable to your purposes. 
Regarding sterimol: I am not entirely sure that the version on the github was the most up to date version of the code (I have updated that version to the one attached to this email).  Clearly that script didn't work as-is, so I ended up doing some minor edits on the code and have posted the updates to github.  I have also updated some of the comments in the script and made it so the code generates two output files ('sterimol_values.csv' and 'biteangle_values.csv').  Just a few notes about running it:
1) Make sure you are using python 2.7 (check out conda https://docs.anaconda.com/anaconda/user-guide/tasks/install-packages/) --> conda create -n sterimolenv python=2.7
2) If on a mac, install openbabel using homebrew (if not see the website for instructions), if on other install open babel via website instructions (make sure to include python bindings)
3) The script will target all XYZ files in a given directory (that means you will have to get the geometry optimized structure from you QM code into XYZ format -- this is required by sterimoltools.py -- though you could likely fiddle with it to make it work with gaussian too)4) Make sure to select the metal, whether or not to include hydrides (likely not, so leave as 0), and the distance cutoff (you don't want to include atoms more than a single bond away from the metal center by accident!).  This will go in the sterimolMultipleSterimol.py header. Open the file and change the "user variables section" how you see fit.

The first output file will consist of the sterimol parameters for every metal-ligand(atom) vector.  This means that if you have a bisphosphine you will get sterimol parameters for both phosphoruses bound to the metal center.  I don't expect this is a problem, but you might want to make the print statements more informative for your purposes.  I attempted to capture the salient information in ouput files, but you will have to configure it for what you want.  I updated the code to write to CSV files.  That way at least it can be parsed without much editing. 

I also pushed through the bite angle portion of the code.  For my grad school work all of the files were prepared in the same way so I always knew that the phosphines in the XYZ file were ordered sequentially (i.e. first and second phosphorus were part of the same ligand, etc).  This code will also generate a second output file which permutes and calculates the angle between every ligand of the metal center. 